Update to 3.1.0; use Fragments
Switch to using fragments which are more efficient and better practice
than relying upon activities. Due to this change, increase the version
number to 3.1.0.

This is almost a complete redo of a lot of the stuff, with the
exception of the ASyncTask. Most things now are more efficient and have
checks to prevent some potential bugs. Up and back navigation have
been implemented properly as well.

Change-Id: I106023f35f29ad16fec695e1bfcbe780411d14bc