Material, New Icon, & other fixes
This really should have gone in as two commits, but is too late now.
Anyway, it gives Paste It a material design and it will reset the api
key when the domain is updated. This will force users to set the proper
api key for their domain.

There is also a new icon that fits the material standards a little
better. It's not perfect, so I'll do some more work on it before making
this final.

Due to the large UI changes, boost the version number to 4.0.0

Not quite ready for merge just yet.

Change-Id: Id18a58327218136bacb709aaaa020466feb5a17f