Fix a crash where app crashes due to orientation change in settings
Formerly, the app crashed when changing the orientation while in
settings, however, this fixes that. This calls SettingsActivity which
handles the placing of SettingsFragment.

Also, it's a wonder that replacing android.R.id.content instead of
R.id.container didn't cause an issue sooner. That bug is now fixed, too.

Change-Id: I20a6c18e6731f5f40347e5ee2c6b1277ae668c05