Implement auto versionCode increase
Our version code system is a mess. Let's just auto increase. We'll never
make it the over 2 billion versionCodes that are available so ot should
be okay that we're doing it.

Right now, it seems to also run everytime you resync the Gradle config,
it also increases the version code, but that's how it goes.

Change-Id: I82e70d6cb2b59cb5e33c8bce426b36321d020e25