Don't Handle Some Exceptions
In Kotlin, exceptions don't have to be checked. In most scenarios in
our code, the exceptions are just logged, which isn't very useful, and
the same generic code was used almost everywhere.

Since there is no requirement to check exceptions, and since we weren't
doing much with our checking, we may as well remove these sections of
our code.

Change-Id: I7961cc9f7584a6c3694cde87aae21afad83462a7